Renamed Ref to Prop. Introduced MultiProp. Delay some Lua calls like AddUnitClass. Added LocalizedName to ISWBFClass.

+++ UnityProject/Assets/Runtime/Scripts/GameRuntime.cs
@@ -181,7 +181,7 @@ public class GameRuntime : MonoBehaviour
     }
 
     // TODO
-    public Color GetTeamColor(byte teamId)
+    public Color GetTeamColor(int teamId)
     {
         switch (teamId)
         {
